pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terml"
version = "0.4.0"
authors = [
    {name = "Paradise Labs", email = "developers@paradiselabs.co"},
]
description = "TerML - AI-powered Terminal Assistant"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "anthropic==0.7.4",
    "click==8.1.3",
    "python-dotenv==1.0.0",
    "pytest==7.3.1",
    "httpx>=0.23.0,<0.28.0",
    "httpcore>=0.15.0,<1.1.0",
]

[project.urls]
"Homepage" = "https://github.com/paradiselabs-ai/TerML"

[project.scripts]
terml = "terml.main:cli"

[tool.setuptools.packages.find]
include = ["terml*"]

[tool.setuptools.package-data]
terml = ["config.py"]